requirement.txt
asgiref==3.5.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
crypto==1.4.1
cryptocode==0.1
cssutils==2.5.0
DateTime==4.5
Django==4.0.2
django-cors-headers==3.11.0
django-crontab==0.7.1
django-environ==0.8.1
django-inlinecss==0.3.0
django-naomi==0.8
djangorestframework==3.13.1
djangorestframework-link-header-pagination==0.1.1
djangorestframework-simplejwt==5.0.0
drf-extra-fields==3.4.0
ecdsa==0.17.0
future==0.18.2
geographiclib==1.52
geopy==2.2.0
idna==3.3
inner==1.1.0
itermate==1.0.2
Naked==0.1.31
numpy==1.22.2
pandas==1.4.1
Pillow==9.0.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycrypto==2.6.1
pycryptodomex==3.14.1
PyJWT==2.3.0
pynliner==0.8.0
PyPDF2==2.4.0
pypi-xmlrpc==2020.12.3
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
qrcode==7.3.1
requests==2.27.1
shellescape==3.8.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
suds==1.1.1
toml==0.10.2
typing_extensions==4.2.0
urllib3==1.26.8
xmltodict==0.12.0
zope.interface==5.4.0
